moves surface related functions (almost unused) from sm_utils to surface_utils, cleanup code
