+++ smtk/sm_utils.py
@@ -25,10 +25,7 @@ import re
 import numpy as np
 from scipy.integrate import cumtrapz
 from scipy.constants import g
-from collections import OrderedDict
-from openquake.hazardlib.geo import (PlanarSurface, SimpleFaultSurface,
-                                     ComplexFaultSurface, MultiSurface,
-                                     Point, Line)
+
 from openquake.hazardlib.gsim import get_available_gsims
 from openquake.hazardlib.scalerel.peer import PeerMSR
 from openquake.hazardlib.gsim.gmpe_table import GMPETable
@@ -105,8 +102,7 @@ def get_time_vector(time_step, number_steps):
     """
     General SMTK utils
     """
-    return np.cumsum(time_step * np.ones(number_steps, dtype=float)) -\
-        time_step
+    return np.cumsum(time_step * np.ones(number_steps, dtype=float)) - time_step
 
 
 def nextpow2(nval):
@@ -180,21 +176,22 @@ def get_velocity_displacement(time_step, acceleration, units="cm/s/s",
 def _save_image(filename, fig, format='png', dpi=300, **kwargs):  # noqa
     """
     Saves the matplotlib figure `fig` to `filename`. Wrapper around `fig.savefig`
-    setting `dpi=300` by default and `format` inferred from `filename` extension
+    with `dpi=300` by default and `format` inferred from `filename` extension
     or, if no extension is found, set as "png".
     If filename is empty this function does nothing and return
 
     :param str filename: str, the file path
     :param figure: a :class:`matplotlib.figure.Figure` (e.g. via
         `matplotlib.pyplot.figure()`)
-    :param format: string, the image format. Default: infer from filename, otherwise
-        set as 'png'
+    :param format: string, the image format. Default: 'png'. This argument is
+        ignored if `filename` has a file extension, as `format` will be set
+        equal to the extension without leading dot.
     :param str kwargs: additional keyword arguments to pass to `fig.savefig`.
         For details, see:
         https://matplotlib.org/stable/api/_as_gen/matplotlib.pyplot.savefig.html
     """
     if not filename:
-        pass
+        return
 
     name, ext = os.path.splitext(filename)
     if ext:
@@ -221,200 +218,52 @@ def load_pickle(pickle_file):
     return pickle_data
 
 
-# Set of modules for converting OpenQuake surface classes to dictionaries
-def planar_fault_surface_to_dict(surface):
-    """
-    Parses a PlanarSurface object to a dictionary formatted for json export
-    """
-    output = OrderedDict([("type", "PlanarSurface")])
-    # Top left
-    for i, key in zip([0, 1, 3, 2],
-        ["top_left", "top_right", "bottom_right", "bottom_left"]):
-        output[key] = [surface.corner_lons[i],
-                       surface.corner_lats[i],
-                       surface.corner_depths[i]]
-    return output
-
-
-def simple_fault_surface_to_dict(surface):
-    """
-    Parsers a SimpleFaultSurface object ot a dictionary formatted for
-    json export
-
-    More complicated here as we need to use the surface nodes
-    """
-    output = OrderedDict([("type", "SimpleFaultSurface")])
-    for node in surface.surface_nodes[0]:
-        if "LineString" in node.tag:
-            trace = node[0].text
-            output["trace"] = [[trace[i], trace[i + 1]]
-                                for i in range(0, len(trace), 2)]
-        else:
-            output[node.tag] = node.text
-    return output
-
-
-def complex_fault_surface_to_dict(surface):
-    """
-    Parses a ComplexFaultSurface object ot a dictionary formatted for json
-    export
-    """
-    output = OrderedDict([("type", "ComplexFaultSurface")])
-    intermediate_edges = []
-    for node in surface.surface_nodes[0]:
-        edge = node[0].nodes[0].text
-        edge = [[edge[i], edge[i + 1], edge[i + 2]]
-                for i in range(0, len(edge), 3)]
-        if "intermediateEdge" in node.tag:
-            intermediate_edges.append(edge)
-        else:
-            output[node.tag] = edge
-    output["intermediateEdges"] = intermediate_edges
-    return output
-
-
-surfaces_to_dict = {
-    "PlanarSurface": planar_fault_surface_to_dict,
-    "SimpleFaultSurface": simple_fault_surface_to_dict,
-    "ComplexFaultSurface": complex_fault_surface_to_dict,
-    }
-
-
-def multi_surface_to_dict(surface):
-    """
-    Parses a multi
-    """
-    output = OrderedDict([("type", "MultiSurface"), ("surfaces", [])])
-    for sfc in surface.surfaces:
-        output["surfaces"].append(
-            surfaces_to_dict[sfc.__class__.__name__](sfc))
-    return output
-
-
-surfaces_to_dict["MultiSurface"] = multi_surface_to_dict
-
-
-# Surfaces from json dict
-def planar_fault_surface_from_dict(data, mesh_spacing=1.):
-    """
-    Builds a planar fault surface from the json load
-    """
-    assert "PlanarSurface" in data
-    top_left = Point(data["top_left"][0],
-                     data["top_left"][1],
-                     data["top_left"][2])
-    top_right = Point(data["top_right"][0],
-                      data["top_right"][1],
-                      data["top_right"][2])
-    bottom_left = Point(data["bottom_left"][0],
-                        data["bottom_left"][1],
-                        data["bottom_left"][2])
-    bottom_right = Point(data["bottom_right"][0],
-                         data["bottom_right"][1],
-                         data["bottom_right"][2])
-    return PlanarSurface.from_corner_points(mesh_spacing, top_left, top_right,
-                                            bottom_right, bottom_left)
-
-
-def simple_fault_surface_from_dict(data, mesh_spacing=1.):
-    """
-    Builds a simple fault surface from the json load
-    """
-    assert "SimpleFaultSurface" in data
-    trace = []
-    for lon, lat in data["trace"]:
-        trace.append(Point(lon, lat, 0.0))
-    trace = Line(trace)
-    return SimpleFaultSurface.from_fault_data(trace, data["upperSeismoDepth"],
-                                              data["lowerSeismoDepth"],
-                                              data["dip"], mesh_spacing)
-
-def _3d_line_from_list(vals):
-    """
-    """
-    vertices = []
-    for lon, lat, depth in vals:
-        vertices.append(Point(lon, lat, depth))
-    return Line(vertices)
-
-def complex_fault_surface_from_dict(data, mesh_spacing=1.):
-    """
-    Builds a complex fault surface from the json load
-    """
-    assert "ComplexFaultSurface" in data
-    edges = [_3d_line_from_list(data["faultTopEdge"])]
-    if len(data["intermediateEdges"]):
-        for iedge in data["intermediateEdges"]:
-            edges.append(_3d_line_from_list(iedge))
-    edges.append(_3d_line_from_list(data["faultBottomEdge"]))
-    return ComplexFaultSurface.from_fault_data(edges, mesh_spacing)
-
-
-surfaces_from_dict = {
-    "PlanarSurface": planar_fault_surface_from_dict,
-    "SimpleFaultSurface": simple_fault_surface_from_dict,
-    "ComplexFaultSurface": complex_fault_surface_from_dict}
-
-
-def multi_surface_from_dict(data, mesh_spacing=1.):
-    """
-    Builds a multi-fault surface from the json load
-    """
-    assert "MultiSurface" in data
-    surfaces = []
-    for surface in data["surfaces"]:
-        surfaces.append(surfaces_from_dict[surface["type"]](surface,
-                                                            mesh_spacing))
-    return MultiSurface(surfaces)
-
-
-surfaces_from_dict["MultiSurface"] = multi_surface_from_dict
-
-
 # Moved from sm_database: Mechanism type to Rake conversion:
-MECHANISM_TYPE = {"Normal": -90.0,
-                  "Strike-Slip": 0.0,
-                  "Reverse": 90.0,
-                  "Oblique": 0.0,
-                  "Unknown": 0.0,
-                  "N": -90.0,  # Flatfile conventions
-                  "S": 0.0,
-                  "R": 90.0,
-                  "U": 0.0,
-                  "NF": -90.,  # ESM flatfile conventions
-                  "SS": 0.,
-                  "TF": 90.,
-                  "NS": -45.,  # Normal with strike-slip component
-                  "TS": 45.,  # Reverse with strike-slip component
-                  "O": 0.0
-                  }
+MECHANISM_TYPE = {
+    "Normal": -90.0,
+    "Strike-Slip": 0.0,
+    "Reverse": 90.0,
+    "Oblique": 0.0,
+    "Unknown": 0.0,
+    "N": -90.0,  # Flatfile conventions
+    "S": 0.0,
+    "R": 90.0,
+    "U": 0.0,
+    "NF": -90.,  # ESM flatfile conventions
+    "SS": 0.,
+    "TF": 90.,
+    "NS": -45.,  # Normal with strike-slip component
+    "TS": 45.,  # Reverse with strike-slip component
+    "O": 0.0
+}
 
 
-DIP_TYPE = {"Normal": 60.0,
-            "Strike-Slip": 90.0,
-            "Reverse": 35.0,
-            "Oblique": 60.0,
-            "Unknown": 90.0,
-            "N": 60.0, # Flatfile conventions
-            "S": 90.0,
-            "R": 35.0,
-            "U": 90.0,
-            "NF": 60., # ESM flatfile conventions
-            "SS": 90.,
-            "TF": 35.,
-            "NS": 70., # Normal with strike-slip component
-            "TS": 45., # Reverse with strike-slip component
-            "O": 90.0
-            }
-
-
+DIP_TYPE = {
+    "Normal": 60.0,
+    "Strike-Slip": 90.0,
+    "Reverse": 35.0,
+    "Oblique": 60.0,
+    "Unknown": 90.0,
+    "N": 60.0, # Flatfile conventions
+    "S": 90.0,
+    "R": 35.0,
+    "U": 90.0,
+    "NF": 60., # ESM flatfile conventions
+    "SS": 90.,
+    "TF": 35.,
+    "NS": 70., # Normal with strike-slip component
+    "TS": 45., # Reverse with strike-slip component
+    "O": 90.0
+}
 
 
 # mean utilities (geometric, arithmetic, ...):
-SCALAR_XY = {"Geometric": lambda x, y: np.sqrt(x * y),
-             "Arithmetic": lambda x, y: (x + y) / 2.,
-             "Larger": lambda x, y: np.max(np.array([x, y]), axis=0),
-             "Vectorial": lambda x, y: np.sqrt(x ** 2. + y ** 2.)}
+SCALAR_XY = {
+    "Geometric": lambda x, y: np.sqrt(x * y),
+    "Arithmetic": lambda x, y: (x + y) / 2.,
+    "Larger": lambda x, y: np.max(np.array([x, y]), axis=0),
+    "Vectorial": lambda x, y: np.sqrt(x ** 2. + y ** 2.)
+}
 
 
 DEFAULT_MSR = PeerMSR()
@@ -444,11 +293,3 @@ def get_interpolated_period(target_period, periods, values):
         np.log10(values[lval]) +
         (np.log10(target_period) - np.log10(periods[lval])) * d_y / d_x
         )
-
-# surfaces_to_dict = {
-#    "PlanarSurface": planar_fault_surface_to_dict,
-#    "SimpleFaultSurface": simple_fault_surface_to_dict,
-#    "ComplexFaultSurface": complex_fault_surface_to_dict,
-#    "MultiSurface": multi_surface_to_dict
-#    }
-
